Assignment4/app/backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
docopt==0.6.2
fastapi==0.135.2
google-auth==2.49.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
mysql==0.0.3
mysql-connector-python==9.6.0
mysqlclient==2.2.8
packaging==26.0
pipreqs==0.4.13
protobuf==7.34.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.33.0
setuptools==80.10.2
SQLAlchemy==2.0.48
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
wheel==0.46.3
yarg==0.1.10